+++ RendererCore/FrameBuffer.h
@@ -8,7 +8,8 @@
 class FrameBuffer {
 public:
     FrameBuffer(int h, int w);
-    bool JudgeDepth(int x, int y, float depth);
+    //bool JudgeDepth(int x, int y, float depth);
+    bool JudgeDepth(float x, float y, float depth);
     void SetPixel(int x, int y, Color color);
     bool SaveImage(QString filePath);
     void ClearBuffer(Color color);
